+++ src/sparseml/pytorch/sparsification/__init__.py
@@ -25,7 +25,6 @@ from .distillation import *
 from .info import *
 from .modifier import *
 from .modifier_thinning import *
-from .obcq import *
 from .pruning import *
 from .quantization.modifier_quantization import *
 from .training import *
