+++ src/sparseml/transformers/sparsification/__init__.py
@@ -22,3 +22,4 @@ Hugging Face transformers flows
 from .question_answering import *
 from .trainer import *
 from .training_args import *
+from .obcq import *
